v8.0 fix bugs and improve events_ecal efficiency 	bugs:	fix long long int bug in initial 		fix a bug in main 		fix print precision by setprecision in global 	improve: if its vacancy skip for loop 		 replace powc function by i*i*i*i 		 (call function takes lot time)

+++ kmc_initial.cpp
@@ -104,7 +104,7 @@ void class_initial::read_restart(char name_restart[], long long int &ts_initial,
 	ifstream if_re(name_restart, ios::in);
 	if(!if_re.is_open()) error(1, "(read_restart) the file is not opened!");
 	
-	int timestep;
+	long long int timestep;
 	double time;
 
 	int ntotal;
